rename wrap_event_handlers to intercept_callback_registrations
