+++ src/boot_env/intercept_callback_registrations.lua
